Remove binary files + tweaks

+++ scripts/prboom.sh
@@ -3,10 +3,19 @@
 set -e
 
 SCRIPT_DIR="$(dirname $(which $0))"
-source "${SCRIPT_DIR}/lib/common.sh"
+source "$(realpath "${SCRIPT_DIR}/lib/common.sh")"
 
-SRC_FILE="${SRC_FILES_DIR}/prboom/prboom.wad"
+REPO_URL="https://github.com/libretro/libretro-prboom.git"
+REPO_NAME="libretro-prboom"
+REPO_PATH="${SRC_REPOS_DIR}/${REPO_NAME}"
+
+SRC_FILE="${REPO_PATH}/prboom.wad"
 ARCHIVE_FILE="${OUT_DIR}/PrBoom.zip"
+
+if ! update_src_repo "$REPO_URL" "$REPO_NAME"
+then
+	exit 1
+fi
 
 rm -f "$ARCHIVE_FILE"
 7z a -mx9 "$ARCHIVE_FILE" "$SRC_FILE"
